requirements.txt
# OceanTune AI — Python dependencies
# Pin exact versions for reproducibility

# Core
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.7.1

# Async & HTTP
aiohttp==3.9.5
httpx==0.27.0
requests==2.32.2

# Data & metrics
pandas==2.2.2
numpy==1.26.4

# Storage — MongoDB (async) + DigitalOcean Spaces (S3-compatible)
motor==3.4.0           # async MongoDB driver (wraps pymongo)
pymongo==4.7.2         # required by motor
boto3==1.34.107

# Logging & observability
rich==13.7.1          # pretty console output
structlog==24.1.0     # structured JSON logging

# Benchmarking / server interaction
openai==1.30.1        # used to talk to the vLLM OpenAI-compatible endpoint

# Optimization
scipy==1.13.0         # for Bayesian optimization helpers (Stage 3+)
scikit-learn==1.4.2   # for random search, cross-validation helpers

# Research agent (Stage 4)
beautifulsoup4==4.12.3
lxml==5.2.2
arxiv==2.1.0

# GitHub interaction
PyGithub==2.3.0

# Node server (FastAPI on each GPU droplet)
fastapi==0.111.0
uvicorn==0.30.1

# CLI
click==8.1.7

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
